refactor: 勝者が決まった際のstyle設定の具体をBoardで行っていたが本来Boardの責務ではなかったため, Squareの方で勝者判定に基づきstyleを行うように変更
