+++ src/components/Board/components/square.tsx
@@ -3,10 +3,16 @@ import { SquareProps } from "../../../type";
 export function Square({
   value,
   onSquareClick,
-  style,
+  isWinner,
 }: SquareProps): JSX.Element {
   return (
-    <button className="square" onClick={onSquareClick} style={style}>
+    <button
+      className="square"
+      onClick={onSquareClick}
+      style={
+        isWinner ? { backgroundColor: "pink" } : { backgroundColor: "white" }
+      }
+    >
       {value}
     </button>
   );
